requirements.txt
Flask==3.0.3
Jinja2==3.1.4
python-dotenv==1.1.1
flask-wtf==1.2.0